Rename some files to be compliant with opfab ts files naming rules
