+++ ui/main/src/app/authentication/CodeAuthenticationHandler.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2024, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,7 +8,7 @@
  */
 
 import {Observable} from 'rxjs';
-import {AuthHandler, HttpAuthInfo} from './auth-handler';
+import {AuthHandler, HttpAuthInfo} from './AuthHandler';
 import {HttpHeaders} from '@angular/common/http';
 import {environment} from '@env/environment';
