+++ ui/main/src/app/builtInTemplates/message-or-question-list/card/MessageOrQuestionListCardTemplateView.ts
