+++ ui/main/src/app/builtInTemplates/message-or-question-list/usercard/MessageOrQuestionListUserCardTemplate.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2024, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,8 +7,8 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {BaseUserCardTemplate} from '../../baseUserCardTemplate';
-import {MessageOrQuestionListUserCardTemplateView} from './message-or-question-listUserCardTemplateView';
+import {BaseUserCardTemplate} from '../../BaseUserCardTemplate';
+import {MessageOrQuestionListUserCardTemplateView} from './MessageOrQuestionListUserCardTemplateView';
 
 declare const opfab;
