+++ ui/main/src/app/builtInTemplates/message/MessageUserCardTemplate.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2024, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,8 +7,8 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {BaseUserCardTemplate} from '../baseUserCardTemplate';
-import {MessageUserCardTemplateView} from './messageUserCardTemplateView';
+import {BaseUserCardTemplate} from '../BaseUserCardTemplate';
+import {MessageUserCardTemplateView} from './MessageUserCardTemplateView';
 
 declare const opfab;
 export class MessageUserCardTemplate extends BaseUserCardTemplate {
