+++ ui/main/src/app/builtInTemplates/question/card/QuestionCardTemplateView.ts
