+++ ui/main/src/app/builtInTemplates/question/usercard/QuestionUserCardTemplateView.spec.ts
@@ -7,7 +7,7 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {QuestionUserCardTemplateView} from './questionUserCardTemplateView';
+import {QuestionUserCardTemplateView} from './QuestionUserCardTemplateView';
 import {initOpfabAPI} from '@tests/helpers';
 import {Card} from 'app/model/Card';
 import {CardTemplateGateway} from '@ofServices/templateGateway/CardTemplateGateway';
