+++ ui/main/src/app/builtInTemplates/task/usercard/TaskUserCardTemplateView.ts
